Refactor RichContract, fix shutdown in makeResource
